+++ src/main/scala/edu/ie3/simona/agent/participant/data/Data.scala
@@ -7,13 +7,13 @@
 package edu.ie3.simona.agent.participant.data
 
 import edu.ie3.datamodel.models.value._
-import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ApparentPower
+import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ComplexPower
 import edu.ie3.util.quantities.PowerSystemUnits
 import edu.ie3.util.quantities.interfaces.EnergyPrice
-import edu.ie3.util.scala.quantities.{Kilovars, Megavars, ReactivePower}
-import squants.energy.{Power, Kilowatts, Megawatts}
-import tech.units.indriya.ComparableQuantity
 import edu.ie3.util.scala.quantities.DefaultQuantities._
+import edu.ie3.util.scala.quantities.{Kilovars, ReactivePower}
+import squants.energy.{Kilowatts, Power}
+import tech.units.indriya.ComparableQuantity
 
 import java.time.ZonedDateTime
 import scala.jdk.OptionConverters.RichOptional
@@ -36,7 +36,7 @@ object Data {
     */
   sealed trait PrimaryData extends Data {
     val p: Power
-    def toApparentPower: ApparentPower
+    def toComplexPower: ComplexPower
   }
 
   object PrimaryData {
@@ -61,7 +61,7 @@ object Data {
       val qDot: Power
     }
 
-    val ZERO_POWER: ApparentPower = ApparentPower(zeroMW, zeroMVAr)
+    val ZERO_POWER: ComplexPower = ComplexPower(zeroMW, zeroMVAr)
 
     /** Active power as participant simulation result
       *
@@ -70,15 +70,15 @@ object Data {
       */
     final case class ActivePower(override val p: Power)
         extends PrimaryData
-        with EnrichableData[ApparentPower] {
-      override def toApparentPower: ApparentPower =
-        ApparentPower(
+        with EnrichableData[ComplexPower] {
+      override def toComplexPower: ComplexPower =
+        ComplexPower(
           p,
           zeroMVAr,
         )
 
-      override def add(q: ReactivePower): ApparentPower =
-        ApparentPower(p, q)
+      override def add(q: ReactivePower): ComplexPower =
+        ComplexPower(p, q)
     }
 
     /** Active and Reactive power as participant simulation result
@@ -88,13 +88,13 @@ object Data {
       * @param q
       *   Reactive power
       */
-    final case class ApparentPower(
+    final case class ComplexPower(
         override val p: Power,
         override val q: ReactivePower,
-    ) extends PrimaryDataWithApparentPower[ApparentPower] {
-      override def toApparentPower: ApparentPower = this
+    ) extends PrimaryDataWithApparentPower[ComplexPower] {
+      override def toComplexPower: ComplexPower = this
 
-      override def withReactivePower(q: ReactivePower): ApparentPower =
+      override def withReactivePower(q: ReactivePower): ComplexPower =
         copy(q = q)
     }
 
@@ -110,15 +110,15 @@ object Data {
         override val qDot: Power,
     ) extends PrimaryData
         with Heat
-        with EnrichableData[ApparentPowerAndHeat] {
-      override def toApparentPower: ApparentPower =
-        ApparentPower(
+        with EnrichableData[ComplexPowerAndHeat] {
+      override def toComplexPower: ComplexPower =
+        ComplexPower(
           p,
           zeroMVAr,
         )
 
-      override def add(q: ReactivePower): ApparentPowerAndHeat =
-        ApparentPowerAndHeat(p, q, qDot)
+      override def add(q: ReactivePower): ComplexPowerAndHeat =
+        ComplexPowerAndHeat(p, q, qDot)
     }
 
     /** Apparent power and heat demand as participant simulation result
@@ -130,16 +130,16 @@ object Data {
       * @param qDot
       *   Heat demand
       */
-    final case class ApparentPowerAndHeat(
+    final case class ComplexPowerAndHeat(
         override val p: Power,
         override val q: ReactivePower,
         override val qDot: Power,
-    ) extends PrimaryDataWithApparentPower[ApparentPowerAndHeat]
+    ) extends PrimaryDataWithApparentPower[ComplexPowerAndHeat]
         with Heat {
-      override def toApparentPower: ApparentPower =
-        ApparentPower(p, q)
+      override def toComplexPower: ComplexPower =
+        ComplexPower(p, q)
 
-      override def withReactivePower(q: ReactivePower): ApparentPowerAndHeat =
+      override def withReactivePower(q: ReactivePower): ComplexPowerAndHeat =
         copy(q = q)
     }
 
@@ -150,7 +150,7 @@ object Data {
             (hs.getP.toScala, hs.getQ.toScala, hs.getHeatDemand.toScala) match {
               case (Some(p), Some(q), Some(qDot)) =>
                 Success(
-                  ApparentPowerAndHeat(
+                  ComplexPowerAndHeat(
                     Kilowatts(
                       p.to(PowerSystemUnits.KILOWATT).getValue.doubleValue
                     ),
@@ -173,7 +173,7 @@ object Data {
             (s.getP.toScala, s.getQ.toScala) match {
               case (Some(p), Some(q)) =>
                 Success(
-                  ApparentPower(
+                  ComplexPower(
                     Kilowatts(
                       p.to(PowerSystemUnits.KILOWATT).getValue.doubleValue
                     ),