+++ src/main/scala/edu/ie3/simona/model/participant/HpModel.scala
@@ -7,7 +7,7 @@
 package edu.ie3.simona.model.participant
 
 import edu.ie3.datamodel.models.input.system.HpInput
-import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ApparentPowerAndHeat
+import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ComplexPowerAndHeat
 import edu.ie3.simona.model.SystemComponent
 import edu.ie3.simona.model.participant.HpModel.{HpRelevantData, HpState}
 import edu.ie3.simona.model.participant.control.QControl
@@ -22,6 +22,7 @@ import edu.ie3.simona.ontology.messages.flex.MinMaxFlexibilityMessage.ProvideMin
 import edu.ie3.util.quantities.PowerSystemUnits
 import edu.ie3.util.scala.OperationInterval
 import edu.ie3.util.scala.quantities.DefaultQuantities._
+import edu.ie3.util.scala.quantities.{ApparentPower, Kilovoltamperes}
 import squants.energy.{KilowattHours, Kilowatts}
 import squants.{Energy, Power, Temperature}
 
@@ -56,13 +57,13 @@ final case class HpModel(
     id: String,
     operationInterval: OperationInterval,
     qControl: QControl,
-    sRated: Power,
+    sRated: ApparentPower,
     cosPhiRated: Double,
     pThermal: Power,
     thermalGrid: ThermalGrid,
 ) extends SystemParticipant[
       HpRelevantData,
-      ApparentPowerAndHeat,
+      ComplexPowerAndHeat,
       HpState,
     ](
       uuid,
@@ -74,8 +75,7 @@ final case class HpModel(
     )
     with ApparentPowerAndHeatParticipant[HpRelevantData, HpState] {
 
-  private val pRated: Power =
-    sRated * cosPhiRated
+  private val pRated: Power = sRated.toActivePower(cosPhiRated)
 
   /** As this is a state-full model (with respect to the current operation
     * condition and inner temperature), the power calculation operates on the
@@ -326,7 +326,7 @@ final case class HpModel(
         updatedHpState.activePower
     val upperBoundary =
       if (canOperate)
-        sRated * cosPhiRated
+        sRated.toActivePower(cosPhiRated)
       else
         zeroKW
 
@@ -361,7 +361,7 @@ final case class HpModel(
       setPower: Power,
   ): (HpState, FlexChangeIndicator) = {
     /* If the set point value is above 50 % of the electrical power, turn on the heat pump otherwise turn it off */
-    val turnOn = setPower > (sRated * cosPhiRated * 0.5)
+    val turnOn = setPower > (sRated.toActivePower(cosPhiRated) * 0.5)
 
     val (
       thermalEnergyDemandHouse,
@@ -430,9 +430,9 @@ object HpModel {
       scaledInput.getId,
       operationInterval,
       qControl,
-      Kilowatts(
+      Kilovoltamperes(
         scaledInput.getType.getsRated
-          .to(PowerSystemUnits.KILOWATT)
+          .to(PowerSystemUnits.KILOVOLTAMPERE)
           .getValue
           .doubleValue
       ),
@@ -533,9 +533,9 @@ object HpModel {
       scaledInput.getId,
       operationInterval,
       qControl,
-      Kilowatts(
+      Kilovoltamperes(
         scaledInput.getType.getsRated
-          .to(PowerSystemUnits.KILOWATT)
+          .to(PowerSystemUnits.KILOVOLTAMPERE)
           .getValue
           .doubleValue
       ),