+++ src/main/scala/edu/ie3/simona/model/participant/evcs/EvcsModel.scala
@@ -11,7 +11,7 @@ import edu.ie3.datamodel.models.ElectricCurrentType
 import edu.ie3.datamodel.models.input.system.EvcsInput
 import edu.ie3.datamodel.models.input.system.`type`.evcslocation.EvcsLocationType
 import edu.ie3.datamodel.models.result.system.{EvResult, EvcsResult}
-import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ApparentPower
+import edu.ie3.simona.agent.participant.data.Data.PrimaryData.ComplexPower
 import edu.ie3.simona.model.SystemComponent
 import edu.ie3.simona.model.participant.control.QControl
 import edu.ie3.simona.model.participant.evcs.EvcsModel._
@@ -32,6 +32,7 @@ import edu.ie3.util.quantities.PowerSystemUnits._
 import edu.ie3.util.quantities.QuantityUtils.RichQuantityDouble
 import edu.ie3.util.scala.OperationInterval
 import edu.ie3.util.scala.quantities.DefaultQuantities._
+import edu.ie3.util.scala.quantities.{ApparentPower, Kilovoltamperes}
 import squants.energy.Kilowatts
 import squants.time.Seconds
 import squants.{Dimensionless, Energy, Power}
@@ -73,7 +74,7 @@ final case class EvcsModel(
     operationInterval: OperationInterval,
     simulationStartDate: ZonedDateTime,
     qControl: QControl,
-    sRated: Power,
+    sRated: ApparentPower,
     currentType: ElectricCurrentType,
     cosPhiRated: Double,
     chargingPoints: Int,
@@ -81,7 +82,7 @@ final case class EvcsModel(
     vehicle2grid: Boolean,
     strategy: ChargingStrategy.Value,
     lowestEvSoc: Double,
-) extends SystemParticipant[EvcsRelevantData, ApparentPower, EvcsState](
+) extends SystemParticipant[EvcsRelevantData, ComplexPower, EvcsState](
       uuid,
       id,
       operationInterval,
@@ -500,12 +501,12 @@ final case class EvcsModel(
   ): Power = {
     val evPower = currentType match {
       case ElectricCurrentType.AC =>
-        ev.sRatedAc
+        ev.pRatedAc
       case ElectricCurrentType.DC =>
-        ev.sRatedDc
+        ev.pRatedDc
     }
     /* Limit the charging power to the minimum of ev's and evcs' permissible power */
-    evPower.min(sRated)
+    evPower.min(sRated.toActivePower(1.0))
   }
 
   override def calculatePower(
@@ -513,7 +514,7 @@ final case class EvcsModel(
       voltage: Dimensionless,
       modelState: EvcsState,
       data: EvcsRelevantData,
-  ): ApparentPower =
+  ): ComplexPower =
     throw new NotImplementedError(
       "Use calculateNewScheduling() or chargeEv() instead."
     )
@@ -750,7 +751,7 @@ final case class EvcsModel(
           if (setPower > zeroKW)
             maxPower
           else
-            maxPower * (-1)
+            maxPower * -1
 
         val chargingTicks = calcFlexOptionsChange(ev, power)
         val endTick = Math.min(currentTick + chargingTicks, ev.departureTick)
@@ -819,7 +820,7 @@ final case class EvcsModel(
 
         (targetEnergy - ev.storedEnergy) / power
       } else
-        (ev.storedEnergy - (ev.eStorage * lowestEvSoc)) / (power * (-1))
+        (ev.storedEnergy - (ev.eStorage * lowestEvSoc)) / (power * -1)
 
     Math.round(timeUntilFullOrEmpty.toSeconds)
   }
@@ -1066,8 +1067,8 @@ object EvcsModel {
       operationInterval,
       simulationStartDate,
       QControl(scaledInput.getqCharacteristics),
-      Kilowatts(
-        scaledInput.getType.getsRated.to(KILOWATT).getValue.doubleValue
+      Kilovoltamperes(
+        scaledInput.getType.getsRated.to(KILOVOLTAMPERE).getValue.doubleValue
       ),
       scaledInput.getType.getElectricCurrentType,
       scaledInput.getCosPhiRated,