+++ src/main/scala/edu/ie3/simona/sim/setup/SetupHelper.scala
@@ -202,17 +202,13 @@ trait SetupHelper extends LazyLogging {
   /** Build the result file hierarchy based on the provided configuration file.
     * The provided type safe config must be able to be parsed as
     * [[SimonaConfig]], otherwise an exception is thrown
+    *
     * @param config
     *   the configuration file
-    * @param createDirs
-    *   if directories of the result file hierarchy should be created or not
     * @return
     *   the resulting result file hierarchy
     */
-  def buildResultFileHierarchy(
-      config: TypesafeConfig,
-      createDirs: Boolean = true,
-  ): ResultFileHierarchy = {
+  def buildResultFileHierarchy(config: TypesafeConfig): ResultFileHierarchy = {
 
     val simonaConfig = SimonaConfig(config)
 
@@ -220,7 +216,7 @@ trait SetupHelper extends LazyLogging {
     val modelsToWrite =
       SetupHelper.allResultEntitiesToWrite(simonaConfig.simona.output)
 
-    val resultFileHierarchy = ResultFileHierarchy(
+    ResultFileHierarchy(
       simonaConfig.simona.output.base.dir,
       simonaConfig.simona.simulationName,
       ResultEntityPathConfig(
@@ -230,15 +226,10 @@ trait SetupHelper extends LazyLogging {
           simonaConfig.simona.simulationName,
         ),
       ),
+      config = Some(config),
       addTimeStampToOutputDir =
         simonaConfig.simona.output.base.addTimestampToOutputDir,
-      createDirs = createDirs,
     )
-
-    // copy config data to output directory
-    ResultFileHierarchy.prepareDirectories(config, resultFileHierarchy)
-
-    resultFileHierarchy
   }
 }