+++ src/test/scala/edu/ie3/simona/model/participant/evcs/uncontrolled/MaximumPowerChargingSpec.scala
@@ -85,7 +85,7 @@ class MaximumPowerChargingSpec
             ScheduleEntry(
               offset,
               offset + expectedDuration,
-              ev.sRatedAc,
+              ev.pRatedAc,
             )
           )
         )